backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==3.2.2
passlib[bcrypt]
pydantic
pydantic[email]
pydantic-settings
python-dotenv
python-multipart